+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/src/CertificateClient.cs
@@ -232,8 +232,39 @@ namespace Azure.Security.KeyVault.Certificates
         /// <exception cref="PlatformNotSupportedException">Cannot create an <see cref="X509Certificate2"/> on this platform.</exception>
         /// <exception cref="RequestFailedException">The request failed. See <see cref="RequestFailedException.ErrorCode"/> and the exception message for details.</exception>
         [CallerShouldAudit(CallerShouldAuditReason)]
-        public virtual Response<X509Certificate2> DownloadCertificate(string certificateName, string version = null, CancellationToken cancellationToken = default) =>
+        [EditorBrowsableAttribute(EditorBrowsableState.Never)]
+        public virtual Response<X509Certificate2> DownloadCertificate(string certificateName, string version, CancellationToken cancellationToken = default) =>
             DownloadCertificate(new DownloadCertificateOptions(certificateName) { Version = version }, cancellationToken);
+
+#pragma warning disable AZC0015 // Unexpected client method return type.
+        /// <summary>
+        /// Creates an <see cref="X509Certificate2"/> from the specified certificate.
+        /// </summary>
+        /// <remarks>
+        /// Because <see cref="KeyVaultCertificate.Cer"/> contains only the public key, this method attempts to download the managed secret
+        /// that contains the full certificate. If you do not have permissions to get the secret,
+        /// <see cref="RequestFailedException"/> will be thrown with an appropriate error response.
+        /// If you want an <see cref="X509Certificate2"/> with only the public key, instantiate it passing only the
+        /// <see cref="KeyVaultCertificate.Cer"/> property.
+        /// This operation requires the certificates/get and secrets/get permissions.
+        /// </remarks>
+        /// <param name="certificateName">The name of the certificate to download.</param>
+        /// <param name="version">(Optional) The version of a certificate to download.</param>
+        /// <param name="outContentType"> The media type (MIME type) in which the certificate will be downloaded. If a supported format is specified,
+        /// the certificate content is converted to the requested format. Currently, only PFX to PEM conversion is supported.
+        /// If an unsupported format is specified, the request is rejected. If not specified, the certificate is returned in its original
+        /// format without conversion. </param>
+        /// <param name="cancellationToken">A <see cref="CancellationToken"/> controlling the request lifetime.</param>
+        /// <returns>An <see cref="X509Certificate2"/> from the specified certificate.</returns>
+        /// <exception cref="ArgumentException"><paramref name="certificateName"/> is empty.</exception>
+        /// <exception cref="ArgumentNullException"><paramref name="certificateName"/> is null.</exception>
+        /// <exception cref="InvalidDataException">The managed secret did not contain a certificate.</exception>
+        /// <exception cref="NotSupportedException">The <see cref="CertificateContentType"/> is not supported.</exception>
+        /// <exception cref="PlatformNotSupportedException">Cannot create an <see cref="X509Certificate2"/> on this platform.</exception>
+        /// <exception cref="RequestFailedException">The request failed. See <see cref="RequestFailedException.ErrorCode"/> and the exception message for details.</exception>
+        [CallerShouldAudit(CallerShouldAuditReason)]
+        public virtual Response<X509Certificate2> DownloadCertificate(string certificateName, string version = null, string outContentType = null, CancellationToken cancellationToken = default) =>
+            DownloadCertificate(new DownloadCertificateOptions(certificateName) { Version = version, OutContentType = outContentType }, cancellationToken);
 
         /// <summary>
         /// Creates an <see cref="X509Certificate2"/> from the specified certificate.
@@ -267,7 +298,16 @@ namespace Azure.Security.KeyVault.Certificates
             try
             {
                 KeyVaultCertificateWithPolicy certificate = _pipeline.SendRequest(RequestMethod.Get, () => new KeyVaultCertificateWithPolicy(), cancellationToken, CertificatesPath, options.CertificateName, "/", options.Version);
-                Response<KeyVaultSecret> secretResponse = _pipeline.SendRequest(RequestMethod.Get, () => new KeyVaultSecret(), certificate.SecretId, appendApiVersion: true, cancellationToken);
+                Response<KeyVaultSecret> secretResponse;
+                if (!string.IsNullOrEmpty(options.OutContentType))
+                {
+                    var requestUri = _pipeline.CompleteUriWithQueryParams(certificate.SecretId, ("outContentType", options.OutContentType));
+                    secretResponse = _pipeline.SendRequest(RequestMethod.Get, () => new KeyVaultSecret(), requestUri, appendApiVersion: false, cancellationToken);
+                }
+                else
+                {
+                    secretResponse = _pipeline.SendRequest(RequestMethod.Get, () => new KeyVaultSecret(), certificate.SecretId, appendApiVersion: true, cancellationToken);
+                }
 
                 KeyVaultSecret secret = secretResponse.Value;
                 string value = secret.Value;
@@ -356,7 +396,16 @@ namespace Azure.Security.KeyVault.Certificates
             try
             {
                 KeyVaultCertificateWithPolicy certificate = await _pipeline.SendRequestAsync(RequestMethod.Get, () => new KeyVaultCertificateWithPolicy(), cancellationToken, CertificatesPath, options.CertificateName, "/", options.Version).ConfigureAwait(false);
-                Response<KeyVaultSecret> secretResponse = await _pipeline.SendRequestAsync(RequestMethod.Get, () => new KeyVaultSecret(), certificate.SecretId, appendApiVersion: true, cancellationToken).ConfigureAwait(false);
+                Response<KeyVaultSecret> secretResponse;
+                if (!string.IsNullOrEmpty(options.OutContentType))
+                {
+                    var requestUri = _pipeline.CompleteUriWithQueryParams(certificate.SecretId, ("outContentType", options.OutContentType));
+                    secretResponse = _pipeline.SendRequest(RequestMethod.Get, () => new KeyVaultSecret(), requestUri, appendApiVersion: false, cancellationToken);
+                }
+                else
+                {
+                    secretResponse = _pipeline.SendRequest(RequestMethod.Get, () => new KeyVaultSecret(), certificate.SecretId, appendApiVersion: true, cancellationToken);
+                }
 
                 KeyVaultSecret secret = secretResponse.Value;
                 string value = secret.Value;